requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
sqlalchemy
psycopg2-binary